backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
openai==1.3.8
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
google-generativeai==0.8.5

